plugin: Make mechanics plugin and shm element
The mechanics plugin is the designated plugin for elements which are
fairly generic "physics". These elements should be the types of stuff
you'd encounter on a physics course in classical mechanics. I think
elastic collisions could be put here as well. I wanted to keep these
distinct from the astro plugin since it has initialisers which are
specific to astrophyics.