refactor!: Calculate the force in integrators
This refactor was done to make implementing RK4 easier. The old
implementation was completely wrong and it required a weird function
that had a "reset" after being called 4 times and each time it did
somthing different. Now integrators can call the force function however
many times they need without giving up control to the pipeline.

The force function is created by the pipeline by chaining together each
individual transform. It is the resposibility of the integrator to call
this function because some integration steps require substeps.
Originally, the substeps were determined using the get_steps method
which the integrator elements were required to implement, but this is
no longer needed.