refactor!: change force to acceleration in core
The core library now handles accelerations rather than forces. The
reason for this is to support different laws of physics as the
original f=ma limits you to classical mechanics. This is a breaking
change since the integrator and transform interfaces have both changed.
The transform elements have been updated to calculate acceleration and
the acceleration calculation has been removed from the integrators. This
change effectively decouples integrators from physics.