perf(astro): iterative get_leaves_with_resolution
By moving from a recursive to an iterative algorithm for the
get_leaves_with_resolution method of QuadTreeNode, I observed 6x
performance improvement. This was due to better memory management since
the the vectors could be pre allocated better. Additionally, the
compiler has a better chance of optimising this method.