feat(cube)!: Add mass to cube galaxy parameters
The mass of the cube galaxy used to be set with the "random" entity.
This was basically a test mode, but it looked cool. The random entity
always had a mass of 0.005, but this isn't useful. Marked as a breaking
change since old simulations will be changed.

BREAKING: cube element has configurable mass (default is 1.0)