feat!: Add fixed field to Entity
Entity now has a boolean fixed field. By default, it is false. If true,
elements can use this field to ignore it. An example of this is keeping
a black hole in the centre of the universe. Before, the message bus and
the entity id were used to implement this behaviour. This is marked as
a break change since entity has a new public field and the ABI is
changed, meaning all plugins must be recompiled with this version of the
library

BREAKING: Added fixed boolean parameter to Entity struct, breaking ABI