feat!: remove unused get/set methods from elements
These methods had no use and needed the user to write a lot of boiler
plate. The messagebus can be used for a similar feature if needed. There
is a chance physim would be used as a library, in which case the get/set
methods would be useful, but I don't think the elements will be used
outside of the main physim program