feat!: add post_creation_messages for transform
After the creation of elements but before the start of the simulation,
elements are given an opportunity to broadcast 1 message. This can be
used to let elements configure each other. A dummy implementation was
added for the astro element and energysink to demonstrate this working.
The side effect of this is harmless for any rendered pipelines, but the
ABI changed, so the major version needs incrementing