feat: graceful handling of panics in transforms
Panics cannot safely unwind across an extern C function call. To prevent
the console getting a largely useless panic unwind, we can catch it with
catch_unwind on the plugin side rather than in the main application.
Most of the time, it makes sense to just abort the program after
printing a useful message about where the problem occurred