fix: Fix memory leak in message bus
Memory was being leaked at the FFI when the recv_message function was
program. This was because of into_raw being called on some strings. This
was fixed by ensuring that from_raw was called *once* on them after the
plugin had finished using them. This was only a problem in the transform
elements though since the recv_message fn took ownership of strings and
none of the other element kinds required CMessage. Now, the recv_message
function borrows the message (instead of mut borrow). This makes sense
since Messages should be read only. I also made the alloc function used
for other methods in the C ABI which allocate strings