feat!: Provide method for C plugin to post to bus
The CMessage struct now takes an enum to determine if the message
originated in Rust of C. This is so the C plugins don't need to use an
allocator to add stuff to the heap. In this way, both C and Rust
messages have their memory managed properly. I also removed unwraps from
the message bus. It will now exit instead of panicking.