fix: deadlock in message bus initialisation sequence
when elements post messages to the message bus, the must aquire a lock
on it to put things on the message queue. The problem was we locked the
bus to add the "pipeline" client, and then tried to lock it again when
we broadcasted the initial configuration messages without dropping the
first lock.