feat!:remove UniverseConfiguration
The render API no longer requires this struct to be passed in. It has
been removed from the core library. It was only required for configuring
the glrender plugin, and all functionality that you'd want is provided
via other configuration option. This was literally just to configure the
view-port and these basically have been constants since day 1 of this
project.